test(cloud): fix configs after MetricPushConcurrency validation
Use NewConfig() defaults (and disable traces where needed) in tests that
construct cloudv2 via New/startVersionedOutput so they keep positive
push concurrency and batch size.

+++ output/cloud/expv2/flush_test.go
@@ -453,7 +453,7 @@ func TestMetricsFlusherRejectsNonPositiveConcurrency(t *testing.T) {
 			case err := <-done:
 				require.Error(t, err)
 				assert.Contains(t, err.Error(), "batch push concurrency must be a positive number")
-				assert.Equal(t, int64(0), pm.timesCalled())
+				assert.Equal(t, 0, pm.timesCalled())
 			case <-time.After(2 * time.Second):
 				t.Fatal("flush hung with non-positive batchPushConcurrency")
 			}
